Use Context to access the instance of the Spotify API instead of using a loader
