+++ src/pages/PlayListPage.jsx
@@ -1,4 +1,4 @@
-import { useLoaderData, useRouteLoaderData } from "react-router-dom";
+import { useLoaderData } from "react-router-dom";
 import List from "@mui/material/List";
 import ListItem from "@mui/material/ListItem";
 import ListItemButton from "@mui/material/ListItemButton";
@@ -7,16 +7,14 @@ import ListItemAvatar from "@mui/material/ListItemAvatar";
 import Avatar from "@mui/material/Avatar";
 import { useNavigate, useParams } from "react-router-dom";
 import { Button } from "@mui/material";
-import { useState } from "react";
-import { useContext } from "react";
-import { AppContext } from "../components/AppContext";
+import { useState, useContext } from "react";
+import { AppContext } from "../App";
 
 export default function PlayListPage() {
   const songs = useLoaderData();
   const navigate = useNavigate();
-  const spotifyApi = useRouteLoaderData("root");
   const { playlistId } = useParams();
-  const { playSelectedSong } = useContext(AppContext);
+  const { playSelectedSong, spotifyApi } = useContext(AppContext);
 
   const [songsOnPlayList, setSongsOnPlayList] = useState(songs);
